refactor(api): move file replace from public API into the plugin files API
`replaceFileFromUploadcareFile` lived on the public UploaderPublicApi, but a
replace (remove + re-add at position, yielding a new internalId) is an advanced,
plugin-oriented mutation — the same category as the existing plugin-only
`filesApi.update`. Consolidate it there and keep the public API focused on
add/remove.

- Move the full implementation into `pluginApi.files.replace`; it now accepts
  `ApiAddFileCommonOptions` (silent/fileName/source) so callers can attribute
  the replacement (e.g. set `source`) like the add* methods.
- Remove `UploaderPublicApi.replaceFileFromUploadcareFile`.
- e2e already drove `pluginApi.files.replace`; extended it to assert the
  `source` option is honored.
