+++ src/abstract/UploaderPublicApi.ts
@@ -142,50 +142,6 @@ export class UploaderPublicApi extends SharedInstance {
     return this.getOutputItem(internalId);
   };
 
-  /**
-   * Replace an existing entry with an already-uploaded Uploadcare file, keeping
-   * the file's position in the list. The original entry is removed and a fresh
-   * entry is added in its place, so the replacement runs through the full add
-   * pipeline (validators, events) like any newly added file. The original's
-   * context (source, metadata, fullPath, silent) is preserved.
-   *
-   * Note: the replacement is a new entry, so it has a NEW `internalId` — use the
-   * returned entry to reference it going forward.
-   */
-  public replaceFileFromUploadcareFile = (
-    internalId: string,
-    file: UploadcareFile,
-    { silent, fileName, source }: ApiAddFileCommonOptions = {},
-  ): OutputFileEntry<'success'> => {
-    const oldId = internalId as Uid;
-    const index = this._uploadCollection.items().indexOf(oldId);
-    const oldEntry = this._uploadCollection.read(oldId);
-    if (index === -1 || !oldEntry) {
-      throw new Error(`File with internalId ${internalId} not found`);
-    }
-    // Carry over the original entry's context so the replacement keeps its
-    // identity within the session; everything file-specific comes from `file`.
-    const preserved = {
-      source: oldEntry.getValue('source'),
-      metadata: oldEntry.getValue('metadata'),
-      fullPath: oldEntry.getValue('fullPath'),
-      silent: oldEntry.getValue('silent'),
-    };
-    this._uploadCollection.remove(oldId);
-    const newId = this._uploadCollection.add(
-      {
-        ...uploadcareFileToEntryData(file),
-        fileName: fileName ?? file.originalFilename ?? null,
-        silent: silent ?? preserved.silent,
-        source: source ?? preserved.source ?? UploadSource.API,
-        metadata: preserved.metadata,
-        fullPath: preserved.fullPath,
-      },
-      { index },
-    );
-    return this.getOutputItem(newId);
-  };
-
   public removeFileByInternalId = (internalId: string): void => {
     if (!this._uploadCollection.read(internalId as Uid)) {
       throw new Error(`File with internalId ${internalId} not found`);
